fix: timezone issue with ISO dates
If the date is in YYYY-MM-DD format, which is the default for the Obsidian 'date' file property, it
will be interpreted as UTC midnight which will mean it renders incorrectly as the previous day when
passed to 'getLocaleDateString'. Adding a time makes it interpret as midnight in the local timezone
instead so it will render as expected.